Make code intel related packages lazy loaded with the new no-op controller
