+++ client/shared/src/codeintel/api.ts
@@ -10,15 +10,19 @@ import {
     TextDocumentIdentifier,
     TextDocumentPositionParameters,
 } from '@sourcegraph/client-api'
+import { MaybeLoadingResult } from '@sourcegraph/codeintellify'
 // eslint-disable-next-line no-restricted-imports
 import { isDefined } from '@sourcegraph/common/src/types'
 import * as clientType from '@sourcegraph/extension-api-types'
 
 import { match } from '../api/client/types/textDocument'
+import { FlatExtensionHostAPI } from '../api/contract'
+import { proxySubscribable } from '../api/extension/api/common'
 import { toPosition } from '../api/extension/api/types'
 import { getModeFromPath } from '../languages'
 import { parseRepoURI } from '../util/url'
 
+import { CodeIntelContext } from './legacy-extensions/api'
 import * as sourcegraph from './legacy-extensions/api'
 import { LanguageSpec } from './legacy-extensions/language-specs/language-spec'
 import { languageSpecs } from './legacy-extensions/language-specs/languages'
@@ -154,4 +158,70 @@ function selectorForSpec(languageSpec: LanguageSpec): DocumentSelector {
         ...(languageSpec.verbatimFilenames || []).flatMap(filename => [{ pattern: filename }]),
         ...languageSpec.fileExts.flatMap(extension => [{ pattern: `*.${extension}` }]),
     ]
+}
+
+// Replaces codeintel functions from the "old" extension/webworker extension API
+// with new implementations of code that lives in this repository. The old
+// implementation invoked codeintel functions via webworkers, and the codeintel
+// implementation lived in a separate repository
+// https://github.com/sourcegraph/code-intel-extensions Ideally, we should
+// update all the usages of `comlink.Remote<FlatExtensionHostAPI>` with the new
+// `CodeIntelAPI` interfaces, but that would require refactoring a lot of files.
+// To minimize the risk of breaking changes caused by the deprecation of
+// extensions, we monkey patch the old implementation with new implementations.
+// The benefit of monkey patching is that we can optionally disable if for
+// customers that choose to enable the legacy extensions.
+export function injectNewCodeintel(
+    old: FlatExtensionHostAPI,
+    codeintelContext: CodeIntelContext
+): FlatExtensionHostAPI {
+    const codeintel = newCodeIntelAPI(codeintelContext)
+    function thenMaybeLoadingResult<T>(promise: Observable<T>): Observable<MaybeLoadingResult<T>> {
+        return promise.pipe(
+            map(result => {
+                const maybeLoadingResult: MaybeLoadingResult<T> = { isLoading: false, result }
+                return maybeLoadingResult
+            })
+        )
+    }
+
+    const codeintelOverrides: Pick<
+        FlatExtensionHostAPI,
+        | 'getHover'
+        | 'getDocumentHighlights'
+        | 'getReferences'
+        | 'getDefinition'
+        | 'getLocations'
+        | 'hasReferenceProvidersForDocument'
+    > = {
+        hasReferenceProvidersForDocument(textParameters) {
+            return proxySubscribable(codeintel.hasReferenceProvidersForDocument(textParameters))
+        },
+        getLocations(id, parameters) {
+            console.log({ id })
+            return proxySubscribable(thenMaybeLoadingResult(codeintel.getImplementations(parameters)))
+        },
+        getDefinition(parameters) {
+            return proxySubscribable(thenMaybeLoadingResult(codeintel.getDefinition(parameters)))
+        },
+        getReferences(parameters, context) {
+            return proxySubscribable(thenMaybeLoadingResult(codeintel.getReferences(parameters, context)))
+        },
+        getDocumentHighlights: (textParameters: TextDocumentPositionParameters) =>
+            proxySubscribable(codeintel.getDocumentHighlights(textParameters)),
+        getHover: (textParameters: TextDocumentPositionParameters) =>
+            proxySubscribable(thenMaybeLoadingResult(codeintel.getHover(textParameters))),
+    }
+
+    return new Proxy(old, {
+        get(target, prop) {
+            // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-explicit-any
+            const codeintelFunction = (codeintelOverrides as any)[prop]
+            if (codeintelFunction) {
+                return codeintelFunction
+            }
+            // eslint-disable-next-line prefer-rest-params
+            return Reflect.get(target, prop, ...arguments)
+        },
+    })
 }
